Lock dialog while network is importing
